refactor: moved checksumprovider from zeebe-db to snapshots
