+++ zeebe/snapshot/src/main/java/io/camunda/zeebe/snapshots/ChecksumProvider.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.db;
+package io.camunda.zeebe.snapshots;
 
 import java.nio.file.Path;
 import java.util.Map;
